refactor: add lib folder for functions
