+++ src/lib/index.js
@@ -1,0 +1,4 @@
+import surveyToText from "./surveyToText.js"
+import { gpt } from "./gptClient.js"
+
+export {gpt, surveyToText}
